enhanced security check.
git-svn-id: https://edk2.svn.sourceforge.net/svnroot/edk2/trunk@7520 6f19259b-4bc3-4df7-8a09-765794883524

+++ edk2/MdeModulePkg/Universal/SetupBrowserDxe/Presentation.c
@@ -231,8 +231,8 @@ DisplayPageFrame (
            Alignment < BANNER_COLUMNS + (UINT8) LocalScreen.LeftColumn;
            Alignment++
           ) {
-        RowIdx = Line - (UINT8) LocalScreen.TopRow;
-        ColumnIdx = Alignment - (UINT8) LocalScreen.LeftColumn;
+        RowIdx = (UINT8) (Line - (UINT8) LocalScreen.TopRow);
+        ColumnIdx = (UINT8) (Alignment - (UINT8) LocalScreen.LeftColumn);
 
         ASSERT (RowIdx < BANNER_HEIGHT);
         ASSERT (ColumnIdx < BANNER_COLUMNS);
